content-loader/pyproject.toml
[project]
name = "content-loader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.7.1",
    "isort>=6.0.1",
    "mcp[cli]>=1.9.3",
    "python-docx>=1.1.2",
    "spire-doc>=13.4.6",
]
